+++ cmake/build.cmake
@@ -2,19 +2,15 @@ cmake_minimum_required(VERSION 3.10)
 
 # display backend type.
 set(DISPLAY_BACKEND_SRC "")
-if(USE_DRM)
-  add_definitions(-DDISPLAY_BACKEND_TYPE_DRM)
+if(${BACKEND_TYPE} STREQUAL "DRM-GBM")
+  add_definitions(-DDISPLAY_BACKEND_TYPE_DRM_GBM)
   set(DISPLAY_BACKEND_SRC
-    src/flutter/shell/platform/linux_embedded/window/linuxes_window_drm.cc
-    src/flutter/shell/platform/linux_embedded/window/native_window_drm.cc
-    src/flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_drm.cc)
-elseif(USE_EGLSTREAM)
-  add_definitions(-DDISPLAY_BACKEND_TYPE_EGLSTREAM)
+    src/flutter/shell/platform/linux_embedded/window/native_window_drm_gbm.cc)
+elseif(${BACKEND_TYPE} STREQUAL "DRM-EGLSTREAM")
+  add_definitions(-DDISPLAY_BACKEND_TYPE_DRM_EGLSTREAM)
   set(DISPLAY_BACKEND_SRC
-    src/flutter/shell/platform/linux_embedded/window/linuxes_window_eglstream.cc
-    src/flutter/shell/platform/linux_embedded/window/native_window_eglstream.cc
-    src/flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_eglstream.cc)
-elseif(USE_X11)
+    src/flutter/shell/platform/linux_embedded/window/native_window_drm_eglstream.cc)
+elseif(${BACKEND_TYPE} STREQUAL "X11")
   add_definitions(-DDISPLAY_BACKEND_TYPE_X11)
   set(DISPLAY_BACKEND_SRC
     src/flutter/shell/platform/linux_embedded/window/linuxes_window_x11.cc
@@ -48,13 +44,13 @@ else()
 endif()
 
 # desktop-shell for weston.
-if(NOT USE_DRM AND DESKTOP_SHELL)
+if(DESKTOP_SHELL)
   add_definitions(-DDESKTOP_SHELL)
 endif()
 
 # wayland & weston protocols.
 set(WAYLAND_PROTOCOL_SRC "")
-if(NOT USE_DRM AND DESKTOP_SHELL)
+if(DESKTOP_SHELL)
   set(WAYLAND_PROTOCOL_SRC ${WAYLAND_PROTOCOL_SRC} src/wayland/protocol/weston-desktop-shell-protocol.c)  
 
   if(USE_VIRTUAL_KEYBOARD)
@@ -160,7 +156,7 @@ target_link_libraries(${TARGET}
     ${USER_APP_LIBRARIES}
 )
 
-if(USE_DRM OR USE_EGLSTREAM)
+if(${BACKEND_TYPE} MATCHES "DRM-(GBM|EGLSTREAM)")
 target_link_libraries(${TARGET}
     PRIVATE
       Threads::Threads
