+++ examples/README.md
@@ -7,7 +7,7 @@ These are an example of how to use embedded Linux embedding for Flutter.
 - [flutter-wayland-client](https://github.com/sony/flutter-embedded-linux/tree/master/examples/flutter-wayland-client): Wayland client app
 - [flutter-weston-desktop-shell](https://github.com/sony/flutter-embedded-linux/tree/master/examples/flutter-weston-desktop-shell): Works on Weston desktop-shell
 - [flutter-weston-desktop-shell-virtual-keyboard](https://github.com/sony/flutter-embedded-linux/tree/master/examples/flutter-weston-desktop-shell-virtual-keyboard): Works on Weston desktop-shell with the virtual keyboard
-- [flutter-drm-backend](https://github.com/sony/flutter-embedded-linux/tree/master/examples/flutter-drm-backend): Fullscreen app on DRM backend
-- [flutter-eglstream-backend](https://github.com/sony/flutter-embedded-linux/tree/master/examples/flutter-eglstream-backend): Fullscreen app on EGLStream backend
+- [flutter-drm-gbm-backend](https://github.com/sony/flutter-embedded-linux/tree/master/examples/flutter-drm-gbm-backend): Fullscreen app on DRM backend with GBM
+- [flutter-drm-eglstream-backend](https://github.com/sony/flutter-embedded-linux/tree/master/examples/flutter-drm-eglstream-backend): Fullscreen app on DRM backend with EGLStream
 - [flutter-x11-client](https://github.com/sony/flutter-embedded-linux/tree/master/examples/flutter-x11-client): X11 client app
 - [flutter-external-texture-plugin](https://github.com/sony/flutter-embedded-linux/tree/master/examples/flutter-external-texture-plugin): Wayland client app using external texture plugin
