+++ examples/flutter-drm-eglstream-backend/README.md
@@ -1,14 +1,14 @@
 # Overview
 
-This is the example of DRM backend.
+This is the example of DRM backend with EGLStream.
 
 ## Building
 
 ```Shell
 $ mkdir build
 $ cd build
-$ cmake -DUSER_PROJECT_PATH=examples/flutter-drm-backend ..
-$ make
+$ cmake -DUSER_PROJECT_PATH=examples/flutter-drm-eglstream-backend ..
+$ cmake --build .
 ```
 
 ## Running Flutter app
@@ -17,7 +17,7 @@ You need to switch from GUI which is running X11 or Wayland to the Character Use
 
 ```Shell
 $ Ctrl + Alt + F3 # Switching to CUI
-$ FLUTTER_DRM_DEVICE="/dev/dri/card1" ./flutter-drm-backend FLUTTER_BUNDLE_PATH
+$ FLUTTER_DRM_DEVICE="/dev/dri/card1" ./flutter-drm-eglstream-backend FLUTTER_BUNDLE_PATH
 ```
 
 Note that replace `FLUTTER_BUNDLE_PATH` with the flutter bundle path you want to use like ./sample/build/linux/x64/release/bundle.
