+++ examples/flutter-external-texture-plugin/cmake/user_config.cmake
@@ -2,9 +2,7 @@ cmake_minimum_required(VERSION 3.10)
 
 # Flutter embedder configurations.
 # See: https://github.com/sony/flutter-embedded-linux#user-configuration-parameters-cmake-options
-set(USE_DRM OFF)
-set(USE_EGLSTREAM OFF)
-set(USE_X11 OFF)
+set(BACKEND_TYPE Wayland)
 set(DESKTOP_SHELL OFF)
 set(USE_VIRTUAL_KEYBOARD OFF)
 set(USE_GLES3 OFF)
