+++ examples/flutter-wayland-client/README.md
@@ -8,5 +8,5 @@ This is the example of Wayland backend + stand-alone application.
 $ mkdir build
 $ cd build
 $ cmake -DUSER_PROJECT_PATH=examples/flutter-wayland-client ..
-$ make
+$ cmake --build .
 ```
