+++ examples/flutter-weston-desktop-shell-virtual-keyboard/README.md
@@ -8,5 +8,5 @@ This is the example of Wayland backend + weston desktop-shell + virtual keyboard
 $ mkdir build
 $ cd build
 $ cmake -DUSER_PROJECT_PATH=examples/flutter-weston-desktop-shell-virtual-keyboard ..
-$ make
+$ cmake --build .
 ```
