+++ examples/flutter-weston-desktop-shell/README.md
@@ -8,5 +8,5 @@ This is the example of Wayland backend + weston desktop-shell.
 $ mkdir build
 $ cd build
 $ cmake -DUSER_PROJECT_PATH=examples/flutter-weston-desktop-shell ..
-$ make
+$ cmake --build .
 ```
