+++ examples/flutter-x11-client/cmake/user_config.cmake
@@ -1,9 +1,7 @@
 cmake_minimum_required(VERSION 3.10)
 
 # Flutter embedder configurations.
-set(USE_DRM OFF)
-set(USE_EGLSTREAM OFF)
-set(USE_X11 ON)
+set(BACKEND_TYPE X11)
 set(DESKTOP_SHELL OFF)
 set(USE_VIRTUAL_KEYBOARD OFF)
 set(USE_GLES3 OFF)
