+++ src/flutter/shell/platform/linux_embedded/surface/context_egl.h
@@ -6,7 +6,6 @@
 #define FLUTTER_SHELL_PLATFORM_LINUX_EMBEDDED_SURFACE_CONTEXT_EGL_H_
 
 #include <EGL/egl.h>
-#include <EGL/eglext.h>
 
 #include <memory>
 
@@ -15,9 +14,6 @@
 #include "flutter/shell/platform/linux_embedded/surface/environment_egl.h"
 #include "flutter/shell/platform/linux_embedded/surface/linuxes_egl_surface.h"
 #include "flutter/shell/platform/linux_embedded/window/native_window.h"
-#if defined(DISPLAY_BACKEND_TYPE_EGLSTREAM)
-#include "flutter/shell/platform/linux_embedded/window/native_window_eglstream.h"
-#endif
 
 namespace flutter {
 
@@ -26,25 +22,8 @@ class ContextEgl {
  public:
   ContextEgl(std::unique_ptr<EnvironmentEgl<D>> environment)
       : environment_(std::move(environment)), config_(nullptr) {
+#if !defined(DISPLAY_BACKEND_TYPE_DRM_EGLSTREAM)
     EGLint config_count = 0;
-#if defined(DISPLAY_BACKEND_TYPE_EGLSTREAM)
-    const EGLint attribs[] = {EGL_SURFACE_TYPE,
-                              EGL_STREAM_BIT_KHR,
-                              EGL_RENDERABLE_TYPE,
-                              EGL_OPENGL_BIT,
-                              EGL_RED_SIZE,
-                              1,
-                              EGL_GREEN_SIZE,
-                              1,
-                              EGL_BLUE_SIZE,
-                              1,
-                              EGL_ALPHA_SIZE,
-                              0,
-                              EGL_DEPTH_SIZE,
-                              1,
-                              EGL_NONE};
-    SetEglExtensionFunctionPointers();
-#else
     const EGLint attribs[] = {
         // clang-format off
         EGL_RENDERABLE_TYPE, EGL_OPENGL_ES2_BIT,
@@ -57,8 +36,6 @@ class ContextEgl {
         EGL_NONE
         // clang-format on
     };
-#endif
-
     if (eglChooseConfig(environment_->Display(), attribs, &config_, 1,
                         &config_count) != EGL_TRUE) {
       LINUXES_LOG(ERROR) << "Failed to choose EGL surface config: "
@@ -91,55 +68,13 @@ class ContextEgl {
     }
 
     valid_ = true;
+#endif
   }
 
   ~ContextEgl() = default;
 
-  std::unique_ptr<LinuxesEGLSurface> CreateOnscreenSurface(
+  virtual std::unique_ptr<LinuxesEGLSurface> CreateOnscreenSurface(
       NativeWindow<W>* window) const {
-#if defined(DISPLAY_BACKEND_TYPE_EGLSTREAM)
-    EGLint layer_count = 0;
-    EGLOutputLayerEXT layer;
-    EGLAttrib layer_attribs[] = {
-        EGL_DRM_PLANE_EXT,
-        static_cast<NativeWindowEglstream*>(window)->PlaneId(),
-        EGL_NONE,
-    };
-
-    if (eglGetOutputLayersEXT_(environment_->Display(), layer_attribs, &layer,
-                               1, &layer_count) != EGL_TRUE) {
-      LINUXES_LOG(ERROR) << "Failed to get EGL output layers";
-      return nullptr;
-    }
-
-    if (layer_count == 0 || layer == nullptr) {
-      LINUXES_LOG(ERROR) << "No matching layers";
-      return nullptr;
-    }
-
-    EGLint stream_attribs[] = {EGL_NONE};
-    auto stream = eglCreateStreamKHR_(environment_->Display(), stream_attribs);
-    if (stream == EGL_NO_STREAM_KHR) {
-      LINUXES_LOG(ERROR) << "Failed to create EGL stream";
-      return nullptr;
-    }
-
-    if (eglStreamConsumerOutputEXT_(environment_->Display(), stream, layer) !=
-        EGL_TRUE) {
-      LINUXES_LOG(ERROR) << "Failed to create EGL stream consumer output";
-      return nullptr;
-    }
-
-    EGLint surface_attribs[] = {
-        EGL_WIDTH, static_cast<NativeWindowEglstream*>(window)->Width(),
-        EGL_HEIGHT, static_cast<NativeWindowEglstream*>(window)->Height(),
-        EGL_NONE};
-    auto surface = eglCreateStreamProducerSurfaceKHR_(
-        environment_->Display(), config_, stream, surface_attribs);
-    if (surface == EGL_NO_SURFACE) {
-      LINUXES_LOG(ERROR) << "Failed to create EGL stream producer surface";
-    }
-#else
     const EGLint attribs[] = {EGL_NONE};
     EGLSurface surface = eglCreateWindowSurface(
         environment_->Display(), config_, window->Window(), attribs);
@@ -147,14 +82,14 @@ class ContextEgl {
       LINUXES_LOG(ERROR) << "Failed to create EGL window surface: "
                          << get_egl_error_cause();
     }
-#endif
     return std::make_unique<LinuxesEGLSurface>(surface, environment_->Display(),
                                                context_);
   }
 
   std::unique_ptr<LinuxesEGLSurface> CreateOffscreenSurface(
       NativeWindow<W>* window_resource) const {
-#if defined(DISPLAY_BACKEND_TYPE_X11) || defined(DISPLAY_BACKEND_TYPE_EGLSTREAM)
+#if defined(DISPLAY_BACKEND_TYPE_X11) || \
+    defined(DISPLAY_BACKEND_TYPE_DRM_EGLSTREAM)
     const EGLint attribs[] = {EGL_WIDTH, 1, EGL_HEIGHT, 1, EGL_NONE};
     EGLSurface surface =
         eglCreatePbufferSurface(environment_->Display(), config_, attribs);
@@ -201,30 +136,12 @@ class ContextEgl {
     return reinterpret_cast<void*>(address);
   }
 
- private:
-  void SetEglExtensionFunctionPointers() {
-    eglGetOutputLayersEXT_ = (PFNEGLGETOUTPUTLAYERSEXTPROC)eglGetProcAddress(
-        "eglGetOutputLayersEXT");
-    eglCreateStreamKHR_ =
-        (PFNEGLCREATESTREAMKHRPROC)eglGetProcAddress("eglCreateStreamKHR");
-    eglStreamConsumerOutputEXT_ =
-        (PFNEGLSTREAMCONSUMEROUTPUTEXTPROC)eglGetProcAddress(
-            "eglStreamConsumerOutputEXT");
-    eglCreateStreamProducerSurfaceKHR_ =
-        (PFNEGLCREATESTREAMPRODUCERSURFACEKHRPROC)eglGetProcAddress(
-            "eglCreateStreamProducerSurfaceKHR");
-  }
-
+ protected:
   std::unique_ptr<EnvironmentEgl<D>> environment_;
   EGLConfig config_;
   EGLContext context_;
   EGLContext resource_context_;
   bool valid_;
-
-  PFNEGLGETOUTPUTLAYERSEXTPROC eglGetOutputLayersEXT_;
-  PFNEGLCREATESTREAMKHRPROC eglCreateStreamKHR_;
-  PFNEGLSTREAMCONSUMEROUTPUTEXTPROC eglStreamConsumerOutputEXT_;
-  PFNEGLCREATESTREAMPRODUCERSURFACEKHRPROC eglCreateStreamProducerSurfaceKHR_;
 };
 
 }  // namespace flutter
