+++ src/flutter/shell/platform/linux_embedded/window_binding_handler.h
@@ -11,10 +11,10 @@
 #include "flutter/shell/platform/linux_embedded/public/flutter_linuxes.h"
 #include "flutter/shell/platform/linux_embedded/window_binding_handler_delegate.h"
 
-#if defined(DISPLAY_BACKEND_TYPE_DRM)
-#include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_drm.h"
-#elif defined(DISPLAY_BACKEND_TYPE_EGLSTREAM)
-#include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_eglstream.h"
+#if defined(DISPLAY_BACKEND_TYPE_DRM_GBM)
+#include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_drm_gbm.h"
+#elif defined(DISPLAY_BACKEND_TYPE_DRM_EGLSTREAM)
+#include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_drm_eglstream.h"
 #elif defined(DISPLAY_BACKEND_TYPE_X11)
 #include "flutter/shell/platform/linux_embedded/surface/linuxes_surface_gl_x11.h"
 #else
@@ -32,10 +32,10 @@ struct PhysicalWindowBounds {
 };
 
 using LinuxesRenderSurfaceTarget =
-#if defined(DISPLAY_BACKEND_TYPE_DRM)
-    SurfaceGlDrm;
-#elif defined(DISPLAY_BACKEND_TYPE_EGLSTREAM)
-    SurfaceGlEglstream;
+#if defined(DISPLAY_BACKEND_TYPE_DRM_GBM)
+    SurfaceGlDrmGbm;
+#elif defined(DISPLAY_BACKEND_TYPE_DRM_EGLSTREAM)
+    SurfaceGlDrmEglstream;
 #elif defined(DISPLAY_BACKEND_TYPE_X11)
     SurfaceGlX11;
 #else
